chore: refactor project to be latest state
Adjust folder structure
