+++ app/novu/workflows/index.ts
@@ -1,0 +1,1 @@
+export * from './welcome-onboarding-email'
