+++ app/novu/workflows/welcome-onboarding-email/index.ts
@@ -1,0 +1,3 @@
+export * from './schemas'
+export * from './types'
+export * from './workflow'
