+++ app/novu/workflows/welcome-onboarding-email/workflow.ts
@@ -1,6 +1,6 @@
 import { workflow } from '@novu/framework';
-import { renderEmail } from './emails/novu-onboarding-email';
-import { zodControlSchema, zodPayloadSchema } from './schemas/definitions';
+import { renderEmail } from '../../emails/novu-onboarding-email';
+import { zodControlSchema, zodPayloadSchema } from './schemas';
 
 export const welcomeOnboardingEmail = workflow(
   'welcome-onboarding-email',
